Port QuicServer to Chrome network stack
Also ported quic_server_bin and quic_server_test.

BUG=

Review URL: https://codereview.chromium.org/340433002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@279602 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/quic/quic_spdy_server_stream.cc
@@ -65,7 +65,7 @@ void QuicSpdyServerStream::OnFinRead() {
   }
 }
 
-int QuicSpdyServerStream::ParseRequestHeaders() {
+size_t QuicSpdyServerStream::ParseRequestHeaders() {
   size_t read_buf_len = static_cast<size_t>(read_buf_->offset());
   SpdyFramer framer(SPDY3);
   SpdyHeaderBlock headers;
